# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/pay/switch-network": {
            "post": {
                "description": "Switch to a different payment network, creating or returning a sub-order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Switch payment network",
                "parameters": [
                    {
                        "description": "Switch network payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SwitchNetworkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/payments/epay/v1/order/create-transaction/submit.php": {
            "get": {
                "description": "Legacy EPAY-style endpoint. Accepts GET (querystring) and POST (form). On success, 302 redirects to /pay/checkout-counter/{trade_id}. Signature uses MD5 of sorted params + secret_key of the api_keys row matching the submitted pid.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Create transaction and redirect (EPAY compat)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key PID (GET query)",
                        "name": "pid",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Amount (fiat, GET query)",
                        "name": "money",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Merchant order ID (GET query)",
                        "name": "out_trade_no",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Callback URL (GET query)",
                        "name": "notify_url",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Redirect URL after payment (GET query)",
                        "name": "return_url",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order name (GET query)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Payment type (e.g. alipay, GET query)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "MD5 signature (GET query)",
                        "name": "sign",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Signature type (MD5, GET query)",
                        "name": "sign_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "API key PID",
                        "name": "pid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Amount (fiat)",
                        "name": "money",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Merchant order ID",
                        "name": "out_trade_no",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Callback URL",
                        "name": "notify_url",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Redirect URL after payment",
                        "name": "return_url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Order name",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Payment type (e.g. alipay)",
                        "name": "type",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "MD5 signature",
                        "name": "sign",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Signature type (MD5)",
                        "name": "sign_type",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to checkout counter"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Legacy EPAY-style endpoint. Accepts GET (querystring) and POST (form). On success, 302 redirects to /pay/checkout-counter/{trade_id}. Signature uses MD5 of sorted params + secret_key of the api_keys row matching the submitted pid.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Create transaction and redirect (EPAY compat)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key PID (GET query)",
                        "name": "pid",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Amount (fiat, GET query)",
                        "name": "money",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Merchant order ID (GET query)",
                        "name": "out_trade_no",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Callback URL (GET query)",
                        "name": "notify_url",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Redirect URL after payment (GET query)",
                        "name": "return_url",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order name (GET query)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Payment type (e.g. alipay, GET query)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "MD5 signature (GET query)",
                        "name": "sign",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Signature type (MD5, GET query)",
                        "name": "sign_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "API key PID",
                        "name": "pid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Amount (fiat)",
                        "name": "money",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Merchant order ID",
                        "name": "out_trade_no",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Callback URL",
                        "name": "notify_url",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Redirect URL after payment",
                        "name": "return_url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Order name",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Payment type (e.g. alipay)",
                        "name": "type",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "MD5 signature",
                        "name": "sign",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Signature type (MD5)",
                        "name": "sign_type",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to checkout counter"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/payments/ezpay/v1/order/create-transaction": {
            "post": {
                "description": "Create a payment transaction order. Accepts JSON body (application/json) or form-encoded body (application/x-www-form-urlencoded).",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Create transaction",
                "parameters": [
                    {
                        "description": "Transaction payload (JSON)",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.CreateTransactionRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Merchant order ID",
                        "name": "order_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Fiat currency (e.g. cny)",
                        "name": "currency",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Crypto token (e.g. usdt)",
                        "name": "token",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Network (e.g. tron)",
                        "name": "network",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "Amount",
                        "name": "amount",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Callback URL",
                        "name": "notify_url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "MD5 signature",
                        "name": "signature",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Redirect URL",
                        "name": "redirect_url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Order name",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Payment type",
                        "name": "payment_type",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.CreateTransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/payments/gmpay/v1/order/create-transaction": {
            "post": {
                "description": "Create a payment transaction order. Accepts JSON body (application/json) or form-encoded body (application/x-www-form-urlencoded).",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Create transaction",
                "parameters": [
                    {
                        "description": "Transaction payload (JSON)",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/request.CreateTransactionRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Merchant order ID",
                        "name": "order_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Fiat currency (e.g. cny)",
                        "name": "currency",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Crypto token (e.g. usdt)",
                        "name": "token",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Network (e.g. tron)",
                        "name": "network",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "Amount",
                        "name": "amount",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Callback URL",
                        "name": "notify_url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "MD5 signature",
                        "name": "signature",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Redirect URL",
                        "name": "redirect_url",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Order name",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Payment type",
                        "name": "payment_type",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.CreateTransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateTransactionRequest": {
            "type": "object",
            "required": [
                "currency",
                "network",
                "notify_url",
                "signature",
                "token"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 100
                },
                "currency": {
                    "description": "法币 如：cny",
                    "type": "string",
                    "example": "cny"
                },
                "name": {
                    "type": "string",
                    "example": "VIP月卡"
                },
                "network": {
                    "description": "网络 如：tron",
                    "type": "string",
                    "example": "tron"
                },
                "notify_url": {
                    "type": "string",
                    "example": "https://example.com/notify"
                },
                "order_id": {
                    "type": "string",
                    "example": "ORD20260416001"
                },
                "payment_type": {
                    "type": "string",
                    "example": "Epay"
                },
                "redirect_url": {
                    "type": "string",
                    "example": "https://example.com/success"
                },
                "signature": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6..."
                },
                "token": {
                    "description": "币种 如：usdt",
                    "type": "string",
                    "example": "usdt"
                }
            }
        },
        "request.SwitchNetworkRequest": {
            "type": "object",
            "required": [
                "network",
                "token",
                "trade_id"
            ],
            "properties": {
                "network": {
                    "type": "string",
                    "example": "ethereum"
                },
                "token": {
                    "type": "string",
                    "example": "USDT"
                },
                "trade_id": {
                    "type": "string",
                    "example": "T2026041612345678"
                }
            }
        },
        "response.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "request_id": {
                    "type": "string",
                    "example": "req-123456"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "response.CreateTransactionResponse": {
            "type": "object",
            "properties": {
                "actual_amount": {
                    "description": "订单实际需要支付的金额，保留4位小数",
                    "type": "number",
                    "example": 14.2857
                },
                "amount": {
                    "description": "订单金额，保留4位小数",
                    "type": "number",
                    "example": 100
                },
                "currency": {
                    "description": "订单货币类型 CNY USD......",
                    "type": "string",
                    "example": "CNY"
                },
                "expiration_time": {
                    "description": "过期时间 时间戳",
                    "type": "integer",
                    "example": 1713264600
                },
                "order_id": {
                    "description": "客户交易id",
                    "type": "string",
                    "example": "ORD20260416001"
                },
                "payment_url": {
                    "description": "收银台地址",
                    "type": "string",
                    "example": "https://pay.example.com/checkout/T2026041612345678"
                },
                "receive_address": {
                    "description": "收款钱包地址",
                    "type": "string",
                    "example": "TTestTronAddress001"
                },
                "token": {
                    "description": "所属币种 TRX USDT......",
                    "type": "string",
                    "example": "USDT"
                },
                "trade_id": {
                    "description": "ezpay订单号",
                    "type": "string",
                    "example": "T2026041612345678"
                }
            }
        },
...
    }
}